+++ sdpDownload/tests/unit/store/mutationsSdpDownload.spec.js
@@ -2,30 +2,9 @@ import {expect} from "chai";
 import mutations from "../../../store/mutationsSdpDownload";
 
 
-const {applyTranslationKey} = mutations;
-
 describe("addons/sdpDownload/store/mutationsSdpDownload", function () {
 
     describe("testing setting of mutations with payload", function () {
-        it("removes 'translate#' from name if present", function () {
-            const state = {
-                    name: "translate#additional:modules.sdpDownload.name"
-                },
-                payload = "translate#additional:modules.sdpDownload.name";
-
-            applyTranslationKey(state, payload);
-            expect(state.name).to.equal("additional:modules.sdpDownload.name");
-        });
-        it("does nothing, if name not starts with 'translate#'", function () {
-            const name = "dies ist ein VueAddon",
-                state = {
-                    name: name
-                },
-                payload = name;
-
-            applyTranslationKey(state, payload);
-            expect(state.name).to.equal(name);
-        });
         it("setSelectedFormat changes value", () => {
             // mock state
             const state = {selectedFormat: "NAS"},
@@ -68,16 +47,4 @@ describe("addons/sdpDownload/store/mutationsSdpDownload", function () {
         });
     });
 
-    describe("testing setting of mutations if payload undefined", function () {
-        it("applyTranslationKey does nothing with undefined payload", function () {
-            const name = "name",
-                state = {
-                    name: name
-                },
-                payload = undefined;
-
-            applyTranslationKey(state, payload);
-            expect(state.name).to.equals(name);
-        });
-    });
 });